+++ src/vm.c
@@ -1287,6 +1287,8 @@ BOOL load_fundamental_classes_on_runtime()
     if(!load_class_with_initialize("Thread")) { return FALSE; }
 
     if(!load_class_with_initialize("Clover")) { return FALSE; }
+    if(!load_class_with_initialize("Null")) { return FALSE; }
+
     gRunningInitializer = FALSE;
 
     return TRUE;