+++ manual/changelog.md
@@ -1,5 +1,15 @@
 English page is here [>> English page](changelog-en)
 
+version 6.6.5
+
+    appended System.nodelay.
+
+version 6.6.4
+
+    appended Null class for REPL
+
+    Fixed some bugs of REPL
+
 version 6.6.3
 
     Fixed block object bug